* commit 'a630ce574':
  msm: camera: reqmgr: Update frame sync logic for dual camera
  msm: camera: isp: Adds register dump to camif irq bottom half
  msm: camera: enables CDM CB for IFE
  msm: camera: ois: Populate delay in I2C register settings
  msm: camera: Dynamically find the request to be deleted
  msm: camera: isp: Multiple fixes in ISP driver
  msm: camera: Memory allocation using cma
  msm: camera: sensor: Fix sensor driver entering infinite loop
  msm: camera: icp: battery drain and hfi setup failure
  msm: camera: smmu: Add functions to map & unmap qdss memory
  msm: camera: reqmgr: Update frame sync logic for dual camera
  msm: camera: icp: Keep AHB vote at SVS

+++ drivers/media/platform/msm/camera/cam_fd/fd_hw_mgr/fd_hw/cam_fd_hw_core.c
@@ -29,10 +29,10 @@ static uint32_t cam_fd_cdm_write_reg_val_pair(uint32_t *buffer,
 }
 
 static void cam_fd_hw_util_cdm_callback(uint32_t handle, void *userdata,
-	enum cam_cdm_cb_status status, uint32_t cookie)
+	enum cam_cdm_cb_status status, uint64_t cookie)
 {
 	trace_cam_cdm_cb("FD", status);
-	CAM_DBG(CAM_FD, "CDM hdl=%x, udata=%pK, status=%d, cookie=%d",
+	CAM_DBG(CAM_FD, "CDM hdl=%x, udata=%pK, status=%d, cookie=%llu",
 		handle, userdata, status, cookie);
 }
